🎈 feat(api/modules/cookie): introduce CookieManager class for cookie management functionality
