+++ apps/api/src/modules/cookie/index.ts
@@ -1,2 +1,1 @@
-export { CookieService } from "./cookie.service";
-export type { ICookieService, CookieAttributes } from "./interfaces/cookie.service.interface";
+export { CookieManager } from "./cookie-manager";
